+++ compile/.fatherrc.ts
@@ -1,0 +1,8 @@
+import { defineConfig } from "father";
+
+export default defineConfig({
+  platform: "node",
+  cjs: {
+    output: "dist",
+  },
+});