+++ templates/mobile/vite.config.ts
@@ -1,3 +1,5 @@
+import pkg from "@vitdoc/compile";
+const { viteIgnore } = pkg;
 import path from "path";
 import { defineConfig } from "vite";
 
@@ -22,4 +24,5 @@ export default defineConfig({
       output: {},
     },
   },
+  plugins: [viteIgnore()],
 });